Converted Knapsack to use ints for indexing only.
The applicable range for the problem implies that the upper 32 bits
for the indexing values are never used anyway. Since the class previously
used a mix of size_t/int values, I made the use of ints consistent.

Reviewed-on: http://gerrit.ent.cloudera.com:8080/3078
Tested-by: jenkins

+++ src/tablet/compaction_policy.cc
@@ -501,7 +501,7 @@ Status BudgetedCompactionPolicy::PickRowSets(const RowSetTree &tree,
   // The value attained by the 'best_chosen' solution.
   double best_optimal = 0;
 
-  vector<size_t> chosen_indexes;
+  vector<int> chosen_indexes;
   vector<CompactionCandidate> inrange_candidates;
   inrange_candidates.reserve(all_candidates.size());
   vector<double> upper_bounds;
